Add files via upload
The primary objective of this project was to leverage data analysis techniques to recommend the most profitable movie genres for production. By diving into the dataset, I sought to uncover insights that would guide decisions on genre selection, thereby enhancing the likelihood of financial success.

NOTE: This was a personal project carried out for my second semester exam.